backend/requirements.txt
# Backend Requirements
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
PyMuPDF==1.23.5
faiss-cpu==1.7.4
transformers==4.33.2
torch==2.0.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
PyJWT==2.8.0
python-dotenv==1.0.0
